+++ api-server-spring-boot-jdbc/src/main/java/com/example/demo/src/user/UserController.java
@@ -21,14 +21,14 @@ public class UserController {
     private final UserProvider userProvider;
     @Autowired
     private final UserService userService;
-    //@Autowired
-    //private final JwtService jwtService;
+    @Autowired
+    private final JwtService jwtService;
 
 
     public UserController(UserProvider userProvider, UserService userService, JwtService jwtService) {
         this.userProvider = userProvider;
         this.userService = userService;
-        //this.jwtService = jwtService;
+        this.jwtService = jwtService;
     }
 
     /**
@@ -58,11 +58,12 @@ public class UserController {
     // Body
     @ResponseBody
     @PostMapping("")
-    public BaseResponse<PostUserRes> createUser(@RequestBody PostUserReq postUserReq) {
-
-
+    public BaseResponse<PostUserRes> createUser(@RequestBody PostUserReq postUserReq) throws BaseException {
         if (postUserReq.getPhoneNumber() == null) {
             return new BaseResponse<>(POST_USERS_EMPTY_PHONENUMBER);
+        }
+        if (postUserReq.getNickName() == null) {
+            return new BaseResponse<>(POST_USERS_EMPTY_NICK_NAME);
         }
         if (!isRegexPhoneNumber(postUserReq.getPhoneNumber())) {
             return new BaseResponse<>(POST_USERS_INVALID_PHONENUMBER);
@@ -77,14 +78,17 @@ public class UserController {
 
     /**
      * 로그인 API
-     * [POST] /users/logIn
+     * [POST] /users/login
      *
      * @return BaseResponse<PostLoginRes>
      */
     @ResponseBody
-    @PostMapping("/logIn")
+    @PostMapping("/login")
     public BaseResponse<PostLoginRes> logIn(@RequestBody PostLoginReq postLoginReq) {
         try {
+            if (postLoginReq.getNickName() == null) {
+                return new BaseResponse<>(POST_USERS_EMPTY_NICK_NAME);
+            }
             if (postLoginReq.getPhoneNumber() == null) {
                 return new BaseResponse<>(POST_USERS_EMPTY_PHONENUMBER);
             }
@@ -109,13 +113,11 @@ public class UserController {
     public BaseResponse<String> modifyUser(@PathVariable("user-id") int userId, @RequestBody User user) {
         try {
             //jwt에서 idx 추출.
-            //int userIdxByJwt = jwtService.getUserIdx();
-            //userIdx와 접근한 유저가 같은지 확인
-            //if (userIdx != userIdxByJwt) {
-            //    return new BaseResponse<>(INVALID_USER_JWT);
-            //}
-
-            //본인 확인 생략!
+            int userIdByJwt = jwtService.getUserId();
+            //userId와 접근한 유저가 같은지 확인
+            if (userId != userIdByJwt) {
+                return new BaseResponse<>(INVALID_USER_JWT);
+            }
             if(user.getNickName() == null){
                 throw new BaseException(PATCH_USERS_EMPTY_NICK_NAME);
             }
@@ -140,13 +142,11 @@ public class UserController {
     public BaseResponse<String> deleteUser(@PathVariable("user-id") int userId) {
         try {
             //jwt에서 idx 추출.
-            //int userIdxByJwt = jwtService.getUserIdx();
+            int userIdByJwt = jwtService.getUserId();
             //userIdx와 접근한 유저가 같은지 확인
-            //if (userIdx != userIdxByJwt) {
-            //    return new BaseResponse<>(INVALID_USER_JWT);
-            //}
-
-            //본인 확인 생략!
+            if (userId != userIdByJwt) {
+                return new BaseResponse<>(INVALID_USER_JWT);
+            }
             userService.deleteUser(userId);
             String result = "";
             return new BaseResponse<>(result);
@@ -154,118 +154,5 @@ public class UserController {
             return new BaseResponse<>((exception.getStatus()));
         }
     }
-
-
-    /*/**
-     * 회원 조회 API
-     * [GET] /users
-     * 회원 번호 및 이메일 검색 조회 API
-     * [GET] /users? Email=
-     * @return BaseResponse<List < GetUserRes>>
-     *//*
-    //Query String
-    @ResponseBody
-    @GetMapping("") // (GET) 127.0.0.1:9000/app/users
-    public BaseResponse<List<GetUserRes>> getUsers(@RequestParam(required = false) String Email) {
-        try{
-            if(Email == null){
-                List<GetUserRes> getUsersRes = userProvider.getUsers();
-                return new BaseResponse<>(getUsersRes);
-            }
-            // Get Users
-            List<GetUserRes> getUsersRes = userProvider.getUsersByEmail(Email);
-            return new BaseResponse<>(getUsersRes);
-        } catch(BaseException exception){
-            return new BaseResponse<>((exception.getStatus()));
-        }
-    }
-
-    /**
-     * 회원 1명 조회 API
-     * [GET] /users/:userIdx
-     * @return BaseResponse<GetUserRes>
-     *//*
-    // Path-variable
-    @ResponseBody
-    @GetMapping("/{userIdx}") // (GET) 127.0.0.1:9000/app/users/:userIdx
-    public BaseResponse<GetUserRes> getUser(@PathVariable("userIdx") int userIdx) {
-        // Get Users
-        try{
-            GetUserRes getUserRes = userProvider.getUser(userIdx);
-            return new BaseResponse<>(getUserRes);
-        } catch(BaseException exception){
-            return new BaseResponse<>((exception.getStatus()));
-        }
-
-    }
-
-    /**
-     * 회원가입 API
-     * [POST] /users
-     * @return BaseResponse<PostUserRes>
-     *//*
-    // Body
-    @ResponseBody
-    @PostMapping("")
-    public BaseResponse<PostUserRes> createUser(@RequestBody PostUserReq postUserReq) {
-        // TODO: email 관련한 짧은 validation 예시입니다. 그 외 더 부가적으로 추가해주세요!
-        if(postUserReq.getEmail() == null){
-            return new BaseResponse<>(POST_USERS_EMPTY_EMAIL);
-        }
-        //이메일 정규표현
-        if(!isRegexEmail(postUserReq.getEmail())){
-            return new BaseResponse<>(POST_USERS_INVALID_EMAIL);
-        }
-        try{
-            PostUserRes postUserRes = userService.createUser(postUserReq);
-            return new BaseResponse<>(postUserRes);
-        } catch(BaseException exception){
-            return new BaseResponse<>((exception.getStatus()));
-        }
-    }
-    /**
-     * 로그인 API
-     * [POST] /users/logIn
-     * @return BaseResponse<PostLoginRes>
-     *//*
-    @ResponseBody
-    @PostMapping("/logIn")
-    public BaseResponse<PostLoginRes> logIn(@RequestBody PostLoginReq postLoginReq){
-        try{
-            // TODO: 로그인 값들에 대한 형식적인 validatin 처리해주셔야합니다!
-            // TODO: 유저의 status ex) 비활성화된 유저, 탈퇴한 유저 등을 관리해주고 있다면 해당 부분에 대한 validation 처리도 해주셔야합니다.
-            PostLoginRes postLoginRes = userProvider.logIn(postLoginReq);
-            return new BaseResponse<>(postLoginRes);
-        } catch (BaseException exception){
-            return new BaseResponse<>(exception.getStatus());
-        }
-    }
-
-    /**
-     * 유저정보변경 API
-     * [PATCH] /users/:userIdx
-     * @return BaseResponse<String>
-     *//*
-    @ResponseBody
-    @PatchMapping("/{userIdx}")
-    public BaseResponse<String> modifyUserName(@PathVariable("userIdx") int userIdx, @RequestBody User user){
-        try {
-            //jwt에서 idx 추출.
-            int userIdxByJwt = jwtService.getUserIdx();
-            //userIdx와 접근한 유저가 같은지 확인
-            if(userIdx != userIdxByJwt){
-                return new BaseResponse<>(INVALID_USER_JWT);
-            }
-            //같다면 유저네임 변경
-            PatchUserReq patchUserReq = new PatchUserReq(userIdx,user.getUserName());
-            userService.modifyUserName(patchUserReq);
-
-            String result = "";
-            return new BaseResponse<>(result);
-        } catch (BaseException exception) {
-            return new BaseResponse<>((exception.getStatus()));
-        }
-    }*/
-
 
 }
